cloud-clone/requirements.txt
colorlog==6.6.0
Flask==2.1.2
Flask_Cors==3.0.10
GitPython==3.1.27
requests==2.27.1
